# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Native Call Center API",
    "version": "1.0.0",
    "summary": "REST and SSE surface of the AI-native call center.",
...
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Same-origin API mount; the SPA is served by the same process."
    }
  ],
...
  "paths": {
...
    "/calls": {
      "get": {
        "operationId": "listCalls",
        "tags": [
          "calls"
        ],
        "summary": "Every live call",
        "description": "The supervision view.",
        "responses": {
          "200": {
            "description": "All live calls.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "cookieSession": [
              "calls:read:all"
            ]
          },
          {
            "apiKeyBearer": [
              "calls:read:all"
            ]
          }
        ]
      },
      "post": {
        "operationId": "createCall",
        "tags": [
          "calls"
        ],
        "summary": "Place a call",
        "description": "One entry point for every call the platform places, told apart by kind.\n\nAI_OUTBOUND originates the customer leg with the DID as caller id; on answer the call bridges to the bot running the DID's flow. AGENT_OUTBOUND is click-to-dial: an agent's own phone is raised first and the destination is dialled when that leg answers. A signed-in agent may omit extensionNumber and place it from the phone they are signed in at; a supervisor or an API-key caller must name the extension, and the phone at it has to be registered.\n\nIdempotent by client-minted callId either way: a retry with the same id answers isDuplicate instead of redialing.",
        "security": [
          {
            "cookieSession": [
              "calls:create"
            ],
            "csrfHeader": []
          },
          {
            "apiKeyBearer": [
              "calls:create"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCallRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The callId was already placed; the response points at the existing call with isDuplicate true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCallResponse"
                }
              }
            }
          },
          "201": {
            "description": "The call is being placed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCallResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aicc_session",
        "description": "Session cookie issued by POST /auth/login, for a person signed in at a browser. The cookie name is configurable (AICC_SESSION_COOKIE); aicc_session is the default. HttpOnly, SameSite=Lax. The scopes it carries are derived from the account's role at sign-in; the role itself never appears on the wire."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AICC-Csrf",
        "description": "Anti-CSRF header, required on every mutating request that authenticates with the session cookie. Any non-empty value passes; requiring a custom header forces a CORS preflight on a cross-origin caller. Missing header: 403 FORBIDDEN. It never accompanies apiKeyBearer — this header protects a credential the browser sends automatically, and a key is sent deliberately."
      },
      "apiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key, presented as `Authorization: Bearer <key>`, for a system with no browser. The key is the subject; it is not a person and never becomes one. To act for an agent — a CRM dialling for somebody on the floor — send `X-AICC-Agent-ID: <agentId>` alongside it. Without that header the subject acts for nobody, and an agent-scoped operation answers AGENT_REQUIRED. The header is refused outright on a session request (AGENT_IMPERSONATION_NOT_ALLOWED), whatever the account's role. A key carries exactly the scopes it was issued with; missing one is INSUFFICIENT_SCOPE."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is malformed: unparseable body or an invalid identifier. Codes VALIDATION_FAILED, USER_DATA_TOO_LARGE.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "No session or the session expired. Codes SESSION_EXPIRED, INVALID_CREDENTIALS.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The identity lacks the role, the CSRF header is missing, the account is suspended, or the account is not an agent. Codes FORBIDDEN, USER_SUSPENDED, NOT_CALL_PARTY.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "The request collides with current state. Codes CONFLICT, EXTENSION_IN_USE, EXTENSION_ASSIGNED_TO_AGENT, EXTENSION_POOL_EXHAUSTED, LAST_ADMIN, AGENT_ALREADY_LOGGED_IN, AGENT_NOT_LOGGED_IN, AGENT_NOT_IN_WRAP_UP.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "BadGateway": {
        "description": "The switch did not take the call. Code SWITCH_DOWN.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "A dependency is unreachable. Codes STORAGE_DOWN, SWITCH_DOWN.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorCode": {
        "type": "string",
        "description": "Machine-readable, translatable failure identifier. The frontend renders errors.<CODE>; the backend never localizes.",
        "enum": [
          "INVALID_CREDENTIALS",
          "SESSION_EXPIRED",
          "FORBIDDEN",
          "AGENT_REQUIRED",
          "AGENT_IMPERSONATION_NOT_ALLOWED",
          "INSUFFICIENT_SCOPE",
          "VALIDATION_FAILED",
          "USER_DATA_TOO_LARGE",
          "NOT_FOUND",
          "METHOD_NOT_ALLOWED",
          "CONFLICT",
          "EXTENSION_IN_USE",
          "EXTENSION_ASSIGNED_TO_AGENT",
          "LAST_ADMIN",
          "EXTENSION_POOL_EXHAUSTED",
          "AGENT_ALREADY_LOGGED_IN",
          "AGENT_NOT_LOGGED_IN",
          "AGENT_NOT_IN_WRAP_UP",
          "CALL_NOT_FOUND",
          "NOT_CALL_PARTY",
          "OPERATION_NOT_ALLOWED_FOR_CALL_TYPE",
          "USER_SUSPENDED",
          "SWITCH_DOWN",
          "STORAGE_DOWN",
          "RATE_LIMITED",
          "INTERNAL"
        ]
      },
      "Error": {
        "type": "object",
        "description": "The single error envelope body: an error code plus interpolation params. Message is diagnostic English, never shown to end users.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "description": "Interpolation parameters for the localized message, e.g. {\"field\": \"extensionNumber\"}.",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CallType": {
        "type": "string",
        "description": "Caller-perspective call type, stamped at creation and immutable across transfers.",
        "enum": [
          "INBOUND",
          "OUTBOUND",
          "CONSULT",
          "INTERNAL"
        ]
      },
      "CallState": {
        "type": "string",
        "enum": [
          "CREATED",
          "RUNNING",
          "ENDED"
        ]
      },
      "PartyRole": {
        "type": "string",
        "enum": [
          "ORIGINATOR",
          "TARGET"
        ]
      },
      "PartyState": {
        "type": "string",
        "enum": [
          "DIALING",
          "RINGING",
          "TALKING",
          "HELD",
          "RELEASED"
        ]
      },
      "PartySnapshot": {
        "type": "object",
        "description": "Immutable view of one call leg.",
        "properties": {
          "partyId": {
            "type": "string",
            "format": "uuid"
          },
          "channelId": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/PartyRole"
          },
          "state": {
            "$ref": "#/components/schemas/PartyState"
          },
          "number": {
            "type": "string"
          },
          "otherNumber": {
            "type": "string"
          },
          "agentId": {
            "type": "string",
            "format": "uuid"
          },
          "isBotLeg": {
            "type": "boolean",
            "description": "True for the leg the switch dialled towards the AI gateway. A caller talking to the bot is on a call with two legs and only one person, and nothing else in this object distinguishes that leg from an agent it has no agentId for."
          },
          "isMuted": {
            "type": "boolean",
            "description": "True while the switch is silencing this leg's microphone."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "answeredAt": {
            "type": "string",
            "format": "date-time"
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "partyId",
          "channelId",
          "role",
          "state",
          "createdAt"
        ]
      },
      "QueueFacts": {
        "type": "object",
        "description": "A call's passage through a queue.",
        "properties": {
          "name": {
            "type": "string"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "bridgedAt": {
            "type": "string",
            "format": "date-time"
          },
          "leftAt": {
            "type": "string",
            "format": "date-time"
          },
          "cause": {
            "type": "string"
          },
          "cancelReason": {
            "type": "string"
          },
          "agentId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CallSnapshot": {
        "type": "object",
        "description": "Immutable view of a live call: the aggregate plus its parties.",
        "properties": {
          "callId": {
            "type": "string",
            "format": "uuid"
          },
          "callType": {
            "$ref": "#/components/schemas/CallType"
          },
          "state": {
            "$ref": "#/components/schemas/CallState"
          },
          "language": {
            "type": "string"
          },
          "queueId": {
            "type": "string",
            "format": "uuid"
          },
          "parties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartySnapshot"
            }
          },
          "userData": {
            "type": "object",
            "description": "Business data attached to the call; merge-patched, survives transfers.",
            "additionalProperties": true
          },
          "queue": {
            "$ref": "#/components/schemas/QueueFacts"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "callId",
          "callType",
          "state",
          "parties",
          "createdAt"
        ]
      },
      "CallList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CallSnapshot"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
      "CreateCallRequest": {
        "type": "object",
        "description": "Ask the platform to place a call. Which fields apply depends on kind.",
        "properties": {
          "callId": {
            "type": "string",
            "format": "uuid",
            "description": "Client-minted id making the request idempotent: a retry with the same id answers isDuplicate instead of redialing."
          },
          "kind": {
            "type": "string",
            "enum": [
              "AI_OUTBOUND",
              "AGENT_OUTBOUND"
            ],
            "description": "AI_OUTBOUND hands the answered customer to the bot; AGENT_OUTBOUND rings an agent's phone first and dials the destination when they pick up. Both produce a call of type OUTBOUND."
          },
          "to": {
            "type": "string",
            "description": "The number to dial."
          },
          "did": {
            "type": "string",
            "description": "AI_OUTBOUND only: the DID whose flow and caller id the call uses."
          },
          "extensionNumber": {
            "type": "string",
            "description": "AGENT_OUTBOUND only: the extension whose phone is raised first. A signed-in agent may leave it out — the phone they signed in at is used — and may not name another's. A supervisor or an API-key caller has no phone of their own, so they must name one."
          },
          "language": {
            "type": "string",
            "description": "AI_OUTBOUND only: overrides the DID's language when set."
          },
          "callbackId": {
            "type": "string",
            "format": "uuid",
            "description": "AGENT_OUTBOUND only: the callback this call is placed to keep. It must be CLAIMED by the caller — a dial for a callback somebody else holds, or nobody does, is refused with 409. The call id is noted on the callback at once and its outcome is copied back when the call ends; the callback itself stays CLAIMED until the agent closes it."
          },
          "userData": {
            "$ref": "#/components/schemas/UserData"
          }
        },
        "required": [
          "kind",
          "to"
        ]
      },
      "CreateCallResponse": {
        "type": "object",
        "properties": {
          "callId": {
            "type": "string",
            "format": "uuid"
          },
          "isDuplicate": {
            "type": "boolean",
            "description": "Present and true when the client-minted callId was already placed; the response points at the existing call."
          }
        },
        "required": [
          "callId"
        ]
      },
...
      "UserData": {
        "type": "object",
        "description": "Business data attached to the call: the order, ticket or case this conversation is about. Flat key/value only, values strings, because this is read as a list of labelled facts and nothing renders a nested object. At most 32 keys, each value at most 1024 bytes of UTF-8; over either limit the request is refused with 400 USER_DATA_TOO_LARGE rather than truncated — a screen showing half a customer's details, with no sign the other half was sent, is worse than a gap. Omitted and {} mean the same thing: no business data. It never reaches the switch.",
        "additionalProperties": {
          "type": "string",
          "maxLength": 1024
        },
        "maxProperties": 32
      },
...
    }
  }
}